Implement send_ping.
For some reason, some responses (namely ping responses without ids)
are being incorrectly returned, with extra null bytes at the end.

The expected response would be:
{"type":"pong"}

Or rather, the bytes:
<<123, 34, 116, 121, 112, 101, 34, 58, 34, 112, 111, 110, 103, 34, 125>>

Instead, I'm receiving the bytes:
<<123, 34, 116, 121, 112, 101, 34, 58, 34, 112, 111, 110, 103, 34, 125,
0, 0, 0, 0, 0, 0>>

I'm unsure if this is a bug in websocket_client or Slack proper. I've
contacted both parties.

For now, we take the response, split it on null bytes, and return
everything before the first null byte.