Fixes mix release deployment
Without this, the application dies with `undef JSX.decode!`.

The fix is described here: https://hexdocs.pm/exrm/extra-common-issues.html

Quoting the page:

Ensure all dependencies for your application are defined in either the `:applications` or `:included_applications` block of your `mix.exs` file. This is how the build process knows that those dependencies need to be bundled in to the release. **This includes dependencies of your dependencies, if they were not properly configured**. For instance, if you depend on `mongoex`, and `mongoex` depends on `erlang-mongodb`, but `mongoex` doesn’t have `erlang-mongodb` in it’s applications list, your app will fail in it’s release form, because `erlang-mongodb` won’t be loaded.