Ensure members is updated when joining a channel
So it would appear that the members section is blank before joining a
channel so if the bot joins the channel after being started it won't
have anything in the `members` key. This appears to fix the problem by
just making sure the `members` key is replaced with the latest `members`
`channel_joined`.