Update slack state before forwarding messages
I'm not 100% sure this is the desired behavior for all use cases but for us at least this is what we want. The main reason this is necessary is that when we receive `:channel_joined` events we are joining a channel for the first time which means we don't have the `members`. If we don't call `handle_slack` first then the `members` key is empty and our `channel_joined` handler doesn't have access to that vital piece of information.