Handle group interactions
Previously, interacting with private groups would fail; Slack does not
have a `group_created` event to parallel its `channel_created` event,
and so the code would crash trying to update data nested under `nil`
whenever a `group_joined` event came in.

To address this, the previously common implementation for channel and
group event handling was split to the extent necessary for dealing with
group interactions. When receiving a `group_joined` event, we simply add
the group map to the `slack` object. There is no reason to keep track of
an `is_member` key, since the user must be a member of the group by
definition.

Users joining and leaving a group is also handled. It turns out that
both a `group_joined` event as well as a `group_join` message are
received, so we ensure that the list of members is maintained
effectively as a set to prevent duplicates.

Finally, when a `group_left` event is received, we simply remove all
record of the group from the `slack` object, since the user is no longer
a member.

Tests included.