Wrap handle_* in `try/rescue`
When an exception occurs in user code happens the output is unhelpful in
diagnosing the problem since there is no stacktrace that can show where
the exception occurred.

This adds `try/rescue` blocks to `handle_message` and `handle_info`
calls so the error message and stacktrace can be printed before
re-raising to crash the genserver.