Fix ignored docs because of missing required key
In the Slack documentation required documentation is specified by a
`required` key in the argument object in the json. Some fields don't
have the key and are ignored since the code was looking for `true` or
`false`.

This fixes the documentation bug by checking for truthy, or non-truthy
values.

This also adds the example to each param description when applicable.