Add workaround for file uploads
Because of the way HTTPoison handles multipart file uploads we have to
treat the upload API endpoint a bit differently.

This adds a workaround where we check if the name of the function is
`upload` and format the parameters for HTTPoison multipart. This works
for now, but if it changes in the future it may make sense to come up
with a more robust solution.