Update to deal with changed presence tracking.
Presence change tracking has changed in slack (see
https://api.slack.com/docs/presence-and-status#batching)

This change updates us to tell slack that we understand the new
presence processing messages, and updates us to handle them properly.

Presence data will still not be supplied to clients until the client
subscribes to presence notifications, so this change also adds a
function to make that a bit more convenient.