Adds the oauth.v2.access method
This implies a breaking change: the current module generator is mapping
multi-level endpoint names to functions with dashes in the name. For example,
`oauth.v2.access` becomes `Slack.Web.Oauth.v2/access/3`.

This, while valid, makes it very dificult to call the function and I believe was
unintended. This PR changes the generator so that the above endpoint becomes
`Slack.Web.Oauth.V2.access/3`.